pyproject.toml
[project]
name = "no-transformer"
version = "0.1.0"
description = "Epistemic multi-hypothesis neuro-symbolic reasoning CLI without transformers"
requires-python = ">=3.11,<3.13"
dependencies = [
    "numpy>=2.0,<3.0",
    "scikit-learn==1.8.0",
    "torch>=2.2,<3.0",
    "sympy>=1.12,<2.0",
    "rank-bm25>=0.2.2,<0.3.0",
    "pyyaml>=6.0,<7.0",
    "prompt_toolkit>=3.0,<4.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0,<9.0"]

[project.scripts]
no-transformer = "cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "ignore::sklearn.exceptions.InconsistentVersionWarning",
]
